requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
python-dotenv==1.0.1
psycopg2-binary==2.9.10
